Revert "move adding participants to use case"
This reverts commit 237afb77d602d4f9abfec340f21df4498aa5b2ba.
