+++ wire-ios-data-model/Source/MLS/Migration/ProteusToMLSMigrationCoordinator.swift
@@ -105,10 +105,7 @@ public class ProteusToMLSMigrationCoordinator: ProteusToMLSMigrationCoordinating
 
         switch migrationStartStatus {
         case .canStart:
-
-            let mlsService = await context.perform { self.context.mlsService }
-
-            guard let mlsService else {
+            guard let mlsService = context.mlsService else {
                 return logger.warn("can't start migration: missing `mlsService`")
             }
