+++ wire-ios/Wire-iOS/Sources/UserInterface/AddContacts/AddParticipantsViewController.swift
@@ -18,7 +18,6 @@
 
 import Foundation
 import UIKit
-import WireSyncEngine
 import WireDataModel
 import WireCommonComponents
 
@@ -101,8 +100,6 @@ final class AddParticipantsViewController: UIViewController, SpinnerCapable {
     }
 
     private let selfUser: UserType
-    private let userSession: UserSession
-
     private let searchResultsViewController: SearchResultsViewController
     private let searchGroupSelector: SearchGroupSelector
     private let searchHeaderViewController: SearchHeaderViewController
@@ -138,14 +135,9 @@ final class AddParticipantsViewController: UIViewController, SpinnerCapable {
 
     convenience init(
         conversation: GroupDetailsConversationType,
-        selfUser: UserType,
-        userSession: UserSession
+        selfUser: UserType
     ) {
-        self.init(
-            context: .add(conversation),
-            selfUser: selfUser,
-            userSession: userSession
-        )
+        self.init(context: .add(conversation), selfUser: selfUser)
     }
 
     override func viewWillDisappear(_ animated: Bool) {
@@ -157,16 +149,11 @@ final class AddParticipantsViewController: UIViewController, SpinnerCapable {
         return wr_supportedInterfaceOrientations
     }
 
-    init(
-        context: Context,
-        selfUser: UserType,
-        userSession: UserSession,
-        isFederationEnabled: Bool = BackendInfo.isFederationEnabled
-    ) {
+    init(context: Context,
+         selfUser: UserType,
+         isFederationEnabled: Bool = BackendInfo.isFederationEnabled) {
 
         self.selfUser = selfUser
-        self.userSession = userSession
-
         viewModel = AddParticipantsViewModel(with: context)
 
         collectionViewLayout = UICollectionViewFlowLayout()
@@ -418,12 +405,7 @@ final class AddParticipantsViewController: UIViewController, SpinnerCapable {
     private func addSelectedParticipants(to conversation: GroupDetailsConversationType) {
         let selectedUsers = self.userSelection.users
 
-        let useCase = AddParticipantsUseCase()
-        useCase.invoke(
-            with: Array(selectedUsers),
-            conversation: conversation,
-            in: userSession
-        )
+        (conversation as? ZMConversation)?.addOrShowError(participants: Array(selectedUsers))
     }
 }
